vulnxscan: Add support for triage
- Add vulnxscan options `--triage` and `--nixprs` to help manual
  analysis.
- Remove nix_secupdates, since all its features are now part of
  vulnxscan.
- Update tests and documentation accordingly.
- Improve repology CVE search
