+++ scripts/vulnxscan/vulnxscan.py
@@ -5,21 +5,32 @@
 # SPDX-License-Identifier: Apache-2.0
 
 # pylint: disable=invalid-name, import-error, too-many-arguments
-# pylint: disable=singleton-comparison
+# pylint: disable=singleton-comparison, abstract-method
+# pylint: disable=too-many-return-statements
 
-""" Demonstrate vulnerability scan with vulnix, grype, and osv.py"""
+"""
+Scan nix artifact or CycloneDX SBOM for vulnerabilities with various
+open-source vulnerability scanners.
+"""
+
 
 import argparse
 import logging
-import os
 import sys
 import pathlib
 import json
 import re
+import time
+import urllib.parse
 from tempfile import NamedTemporaryFile
 from shutil import which
+
+from requests import Session
+from requests_cache import CacheMixin
+from requests_ratelimiter import LimiterMixin
 import pandas as pd
 import numpy as np
+
 from tabulate import tabulate
 from scripts.vulnxscan.osv import OSV
 from sbomnix.sbomdb import SbomDb
@@ -30,8 +41,11 @@ from sbomnix.utils import (
     exec_cmd,
     df_to_csv_file,
     df_from_csv_file,
-    exit_unless_nix_artifact,
     df_log,
+    exit_unless_nix_artifact,
+    nix_to_repology_pkg_name,
+    parse_version,
+    version_distance,
 )
 
 ###############################################################################
@@ -40,10 +54,10 @@ from sbomnix.utils import (
 def getargs():
     """Parse command line arguments"""
     desc = (
-        "Scan nix artifact or CycloneDX SBOM for vulnerabilities with grype, "
-        "osv.py, and vulnix."
+        "Scan nix artifact or CycloneDX SBOM for vulnerabilities with "
+        "various open-source vulnerability scanners."
     )
-    epil = f"Example: ./{os.path.basename(__file__)} /path/to/nix/out/or/drv"
+    epil = "Example: ./vulnxscan.py /path/to/nix/out/or/drv"
     parser = argparse.ArgumentParser(description=desc, epilog=epil)
     helps = "Target derivation path or nix out path"
     parser.add_argument("TARGET", help=helps, type=pathlib.Path)
@@ -73,13 +87,36 @@ def getargs():
     helps = (
         "Path to whitelist file. Vulnerabilities that match any whitelisted "
         "entries will not be included to the console output and are annotated "
-        "accordingly in the output csv."
+        "accordingly in the output csv. See more details in the vulnxscan "
+        "README.md."
     )
     parser.add_argument("--whitelist", help=helps, type=pathlib.Path)
+    helps = (
+        "Add more information to vulnxscan output by querying "
+        "repology.org for available package versions in nix-unstable and "
+        "package upstream. This option is intended to help manual analysis. "
+        "Output is written to a separate OUT file with 'triage' infix, "
+        "by default: 'vulns.triage.csv'."
+    )
+    parser.add_argument("--triage", help=helps, action="store_true")
+    triagegr = parser.add_argument_group("Other arguments")
+    helps = (
+        "Search nixpkgs github for PRs that might include more information "
+        "concerning possible nixpkgs fixes for the found vulnerabilities. "
+        "This option adds URLs to (at most five) PRs that appear valid "
+        "for each vulnerability based on heuristic. "
+        "The PR search takes significant "
+        "time due to github API rate limits, which is why this feature is "
+        "not enabled by default. This option has no impact unless '--triage' "
+        "is also specified."
+    )
+    triagegr.add_argument("--nixprs", help=helps, action="store_true")
     return parser.parse_args()
 
 
 ################################################################################
+
+# Scanner
 
 
 class VulnScan:
@@ -90,6 +127,7 @@ class VulnScan:
         self.df_grype = None
         self.df_osv = None
         self.df_report = None
+        self.df_triaged = None
 
     def _parse_vulnix(self, json_str):
         vulnerable_packages = json.loads(json_str)
@@ -256,8 +294,42 @@ class VulnScan:
             return
         df_apply_whitelist(df_whitelist, self.df_report)
 
+    def _console_report(self):
+        LOG.debug("")
+        # Copy the df to not make changes to the original dataframe
+        if self.df_triaged is not None:
+            df = self.df_triaged.copy()
+            df = df.drop("package_repology", axis=1)
+        else:
+            df = self.df_report.copy()
+        # Don't print the following columns to console
+        df = df.drop("sortcol", axis=1)
+        #  Don't print whitelisted entries to the console report
+        df = df_drop_whitelisted(df)
+        if df.empty:
+            LOG.info("Whitelisted all vulnerabilities")
+            return
+        # Truncate version columns
+        version_cols = [col for col in df.columns if "version" in col]
+        for col in version_cols:
+            df[col] = df[col].str.slice(0, 16)
+        # Write the console report
+        table = tabulate(
+            df,
+            headers="keys",
+            tablefmt="orgtbl",
+            numalign="center",
+            showindex=False,
+        )
+        LOG.info(
+            "Console report\n\n"
+            "Potential vulnerabilities impacting version_local: "
+            "\n\n%s\n\n",
+            table,
+        )
+
     def report(self, args, sbom_csv):
-        """Generate the vulnerability report: csv file and a table to console"""
+        """Generate the vulnerability reports: csv file and a table to console"""
         self._generate_report()
         if self.df_report is None or self.df_report.empty:
             LOG.info("No vulnerabilities found")
@@ -267,43 +339,338 @@ class VulnScan:
         if args.whitelist:
             LOG.info("Applying whitelist '%s'", args.whitelist)
             self._apply_whitelist(args.whitelist)
-        LOG.debug("Writing report")
+        if args.triage:
+            LOG.info("Running vulnerability triage")
+            self.df_triaged = _triage(self.df_report, args.nixprs)
+        # Rename 'version' to 'version_local'
+        self.df_report.rename(columns={"version": "version_local"}, inplace=True)
 
+        LOG.debug("Writing reports")
         # Console report
-        # Copy the df to only make changes to the console report
-        df = self.df_report.copy()
-        # Don't print the "sortcol"
-        df.drop("sortcol", inplace=True, axis=1)
-        # Don't print whitelisted entries to the console report
-        df = df_drop_whitelisted(df)
-        if df.empty:
-            LOG.info("Whitelisted all vulnerabilities")
-        else:
-            # Truncate
-            df["version"] = df["version"].str.slice(0, 16)
-            table = tabulate(
-                df,
-                headers="keys",
-                tablefmt="orgtbl",
-                numalign="center",
-                showindex=False,
-            )
-            if args.sbom:
-                end = f"components in '{args.TARGET}'"
-            elif args.buildtime:
-                end = (
-                    f"'{args.TARGET}' or some of its runtime or buildtime dependencies"
-                )
-            else:
-                end = f"'{args.TARGET}' or some of its runtime dependencies"
-            header = f"Potential vulnerabilities impacting {end}:"
-            LOG.info("Console report\n\n%s\n\n%s\n", header, table)
-
+        self._console_report()
         # File report
-        df_to_csv_file(self.df_report, args.out)
+        out = pathlib.Path(args.out)
+        df_to_csv_file(self.df_report, out.resolve().as_posix())
+        if args.triage:
+            parents = out.parents[0].resolve().as_posix()
+            triage_out = f"{parents}/{out.stem}.triage{out.suffix}"
+            df_to_csv_file(self.df_triaged, triage_out)
 
 
 ################################################################################
+
+# Triage
+
+
+class CachedLimiterSession(CacheMixin, LimiterMixin, Session):
+    """Session class with caching and rate-limiting"""
+
+
+_repology_cve_dfs = {}
+_repology_cli_dfs = {}
+_repology_nix_repo = "nix_unstable"
+# Rate-limited and cached session. For github api rate limits, see:
+# https://docs.github.com/en/rest/search?apiVersion=latest#rate-limit
+_session = CachedLimiterSession(per_minute=9, per_second=1, expire_after=7200)
+
+
+def _select_newest(df):
+    df_ret = None
+    for pkg_name in df["package"].unique():
+        df_pkg = df[df["package"] == str(pkg_name)]
+        # Newest by status:
+        df_newest = df_pkg[df_pkg["status"] == "newest"]
+        if df_newest.empty:
+            # If there's no newest by status, take the newest by version:
+            df_newest = df_pkg.sort_values(by=["version"]).iloc[[-1]]
+        df_ret = pd.concat([df_ret, df_newest], ignore_index=True)
+    return df_ret
+
+
+def _run_repology_cli(pname, match_type="--pkg_exact"):
+    LOG.log(LOG_SPAM, "Running repology_cli for '%s'", pname)
+    df_repology_cli = None
+    if pname in _repology_cli_dfs:
+        LOG.log(LOG_SPAM, "Using cached repology_cli results")
+        df_repology_cli = _repology_cli_dfs[pname]
+    else:
+        prefix = "repology_cli_"
+        suffix = ".csv"
+        with NamedTemporaryFile(delete=True, prefix=prefix, suffix=suffix) as f:
+            repo = f"--repository {_repology_nix_repo}"
+            status = "--re_status=outdated|newest|devel|unique"
+            out = f"--out={f.name}"
+            search = f"{match_type}={pname}"
+            cmd = f"repology_cli.py {repo} {status} {search} {out} "
+            ret = exec_cmd(cmd.split(), raise_on_error=False, return_error=True)
+            if ret and ret.stderr and "No matching packages" in ret.stderr:
+                return None
+            df_repology_cli = df_from_csv_file(f.name)
+            df_repology_cli = _select_newest(df_repology_cli)
+            _repology_cli_dfs[pname] = df_repology_cli
+            df_log(df_repology_cli, LOG_SPAM)
+    return df_repology_cli
+
+
+def _add_vuln_item(out_dict, vuln, whitelist_cols, df_repo=None):
+    if df_repo is None:
+        out_dict.setdefault("vuln_id", []).append(vuln.vuln_id)
+        out_dict.setdefault("url", []).append(vuln.url)
+        out_dict.setdefault("package", []).append(vuln.package)
+        out_dict.setdefault("version_local", []).append(vuln.version)
+        out_dict.setdefault("version_nixpkgs", []).append("")
+        out_dict.setdefault("version_upstream", []).append("")
+        out_dict.setdefault("package_repology", []).append("")
+        out_dict.setdefault("sortcol", []).append(vuln.sortcol)
+        if whitelist_cols:
+            out_dict.setdefault("whitelist", []).append(vuln.whitelist)
+            out_dict.setdefault("whitelist_comment", []).append(vuln.whitelist_comment)
+        return
+    for item in df_repo.itertuples():
+        out_dict.setdefault("vuln_id", []).append(vuln.vuln_id)
+        out_dict.setdefault("url", []).append(vuln.url)
+        out_dict.setdefault("package", []).append(vuln.package)
+        out_dict.setdefault("version_local", []).append(vuln.version)
+        out_dict.setdefault("version_nixpkgs", []).append(item.version)
+        if item.newest_upstream_release and ";" in item.newest_upstream_release:
+            version_upstream_str = item.newest_upstream_release.split(";")[0]
+        else:
+            version_upstream_str = item.newest_upstream_release
+        out_dict.setdefault("version_upstream", []).append(version_upstream_str)
+        out_dict.setdefault("package_repology", []).append(item.package)
+        out_dict.setdefault("sortcol", []).append(vuln.sortcol)
+        if whitelist_cols:
+            out_dict.setdefault("whitelist", []).append(vuln.whitelist)
+            out_dict.setdefault("whitelist_comment", []).append(vuln.whitelist_comment)
+
+
+def _version_similarity(row):
+    ratio = version_distance(row.version, row.version_cmp)
+    LOG.log(
+        LOG_SPAM,
+        "Version similarity ('%s' vs '%s' ==> %s)",
+        row.version,
+        row.version_cmp,
+        ratio,
+    )
+    return ratio
+
+
+def _query_repology_versions(df_vuln_pkgs):
+    LOG.info("Querying repology")
+    result_dict = {}
+    wcols = "whitelist" in df_vuln_pkgs.columns
+    for vuln in df_vuln_pkgs.itertuples():
+        if wcols and vuln.whitelist:
+            # Skip repology query for whitelisted vulnerabilities
+            LOG.log(LOG_SPAM, "Whitelisted, skipping repology query: %s", vuln)
+            _add_vuln_item(result_dict, vuln, wcols)
+            continue
+        repo_pkg = nix_to_repology_pkg_name(vuln.package)
+        LOG.log(LOG_SPAM, "Package '%s' ==> '%s'", vuln.package, repo_pkg)
+        df_repology_cli = _run_repology_cli(repo_pkg)
+        if df_repology_cli is not None and not df_repology_cli.empty:
+            # If there's one match, there's no need to check other details
+            if df_repology_cli.shape[0] == 1:
+                LOG.log(LOG_SPAM, "One repology package matches")
+                _add_vuln_item(result_dict, vuln, wcols, df_repology_cli)
+                continue
+            # Match based on version: exact match
+            df = df_repology_cli[df_repology_cli["version"] == vuln.version]
+            if not df.empty:
+                LOG.log(LOG_SPAM, "Exact version match '%s'", vuln.version)
+                _add_vuln_item(result_dict, vuln, wcols, df)
+                continue
+            # Match based on version: similarity
+            df_repology_cli["version_cmp"] = vuln.version
+            df_repology_cli["similarity"] = df_repology_cli.apply(
+                _version_similarity, axis=1
+            )
+            df = df_repology_cli[df_repology_cli["similarity"] >= 0.7]
+            if not df.empty:
+                LOG.log(LOG_SPAM, "Version similarity match:\n%s", df)
+                best_match = df["similarity"].max()
+                df = df[df["similarity"] == best_match]
+                LOG.log(LOG_SPAM, "Selecting best match based on version:\n%s", df)
+                _add_vuln_item(result_dict, vuln, wcols, df)
+                continue
+            # Otherwise, we need to conclude that we don't know which repology
+            # package (as returned by _run_repology_cli()), the nix package
+            # 'vuln.package' maps to.
+            # If we end up here, we could improve by doing another search with:
+            # _run_repology_cli(repo_pkg, match_type='--pkg_search')
+            LOG.log(LOG_SPAM, "Vague match in repology pkg, adding vuln only")
+            _add_vuln_item(result_dict, vuln, wcols)
+        else:
+            _add_vuln_item(result_dict, vuln, wcols)
+    df_result = pd.DataFrame(result_dict)
+    df_result.fillna("", inplace=True)
+    df_result.reset_index(drop=True, inplace=True)
+    return df_result
+
+
+def _pkg_is_vulnerable(repo_pkg_name, pkg_version, cve_id=None):
+    """
+    Return true if given pkg version is vulnerable. If cve_id is specified,
+    return true only if pkg is affected by the given cve id.
+    """
+    # For now, we rely on repology for the vulnerability info.
+    LOG.debug("Finding vulnerability status for %s:%s", repo_pkg_name, pkg_version)
+    key = f"{repo_pkg_name}:{pkg_version}"
+    if key in _repology_cve_dfs:
+        LOG.log(LOG_SPAM, "Using cached repology_cve results")
+        df = _repology_cve_dfs[key]
+    else:
+        prefix = "repology_cve_"
+        suffix = ".csv"
+        with NamedTemporaryFile(delete=True, prefix=prefix, suffix=suffix) as f:
+            args = f"{repo_pkg_name} {pkg_version}"
+            cmd = f"repology_cve.py --out={f.name} {args}"
+            exec_cmd(cmd.split(), raise_on_error=False)
+            try:
+                df = df_from_csv_file(f.name)
+                df_log(df, LOG_SPAM)
+            except pd.errors.EmptyDataError:
+                df = pd.DataFrame()
+        _repology_cve_dfs[key] = df
+    if cve_id and not df.empty:
+        df = df[df["cve"] == cve_id]
+    return not df.empty
+
+
+def _vuln_update_classify(row):
+    if not row.version_nixpkgs and not row.version_upstream:
+        return "err_missing_repology_version"
+    # Check that the package is also vulnerable based on repology
+    if row.version_local and not _pkg_is_vulnerable(
+        row.package_repology, row.version_local, row.vuln_id
+    ):
+        return "err_not_vulnerable_based_on_repology"
+    # Check if there's an update available in nixpkgs
+    version_local = parse_version(row.version_local)
+    version_nixpkgs = parse_version(row.version_nixpkgs)
+    if not version_local:
+        return "err_invalid_version"
+    if not version_nixpkgs:
+        return "err_invalid_version"
+    if row.version_nixpkgs and version_local < version_nixpkgs:
+        # Classify accordingly if the nixpkgs update is not vulnerable
+        if not _pkg_is_vulnerable(
+            row.package_repology, row.version_nixpkgs, row.vuln_id
+        ):
+            return "fix_update_to_version_nixpkgs"
+    # Check if there's an update available in upstream
+    version_upstream = parse_version(row.version_upstream)
+    if not version_upstream:
+        return "err_invalid_version"
+    if row.version_upstream and version_local < version_upstream:
+        # Classify accordingly if the upstream update is not vulnerable
+        if not _pkg_is_vulnerable(row.package_repology, version_upstream, row.vuln_id):
+            return "fix_update_to_version_upstream"
+    # Issue appears valid, but there's no known fix
+    return "fix_not_available"
+
+
+def _search_result_append(prs, result):
+    maxres = 5
+    for item in prs["items"]:
+        if len(result) >= maxres:
+            # Log all the found PR urls, even though we include only the first
+            # maxres to the main output
+            LOG.log(
+                LOG_SPAM, "More than %s PRs, skipping: %s", maxres, item["html_url"]
+            )
+            continue
+        result.add(item["html_url"])
+    return result
+
+
+def _vuln_nixpkgs_pr(row):
+    if hasattr(row, "whitelist") and row.whitelist:
+        # Whitelisted: skip github nixpkgs pr search
+        LOG.log(LOG_SPAM, "Whitelisted, skipping PR query: %s", row)
+        return ""
+    # See: https://docs.github.com/en/search-github
+    nixpr = "repo:NixOS/nixpkgs is:pr"
+    unmerged = "is:unmerged is:open"
+    merged = "is:merged"
+    ver = None
+    result = set()
+    # Query unmerged PRs based on vuln_id
+    prs = _github_query(f"{nixpr} {unmerged} {row.vuln_id}")
+    _search_result_append(prs, result)
+    # Query merged PRs based vuln_id
+    prs = _github_query(f"{nixpr} {merged} {row.vuln_id}")
+    _search_result_append(prs, result)
+    # Attempt version-based match for the following classifications:
+    if row.classify == "fix_update_to_version_nixpkgs":
+        ver = row.version_nixpkgs
+    elif row.classify == "fix_update_to_version_upstream":
+        ver = row.version_upstream
+    if ver:
+        pkg = row.package
+        # Query unmerged PRs based on pkg name and version in title
+        prs = _github_query(f"{nixpr} {unmerged} {pkg} in:title {ver} in:title")
+        _search_result_append(prs, result)
+        # Query merged PRs based on pkg name and version in title
+        prs = _github_query(f"{nixpr} {merged} {pkg} in:title {ver} in:title")
+        _search_result_append(prs, result)
+    return " \n".join(sorted(list(result)))
+
+
+def _github_query(query_str, delay=60):
+    query_str_quoted = urllib.parse.quote(query_str, safe=":/")
+    query = f"https://api.github.com/search/issues?q={query_str_quoted}"
+    LOG.debug("GET: %s", query)
+    resp = _session.get(query)
+    if not resp.ok and "rate limit exceeded" in resp.text:
+        max_delay = 60
+        if delay > max_delay:
+            LOG.warning("Rate limit exceeded requesting %s", query)
+            ret = json.loads("{}")
+            ret["items"] = []
+            return ret
+        LOG.debug("Sleeping %s seconds before re-requesting", delay)
+        time.sleep(delay)
+        LOG.debug("Re-requesting")
+        return _github_query(query_str, delay * 2)
+    resp.raise_for_status()
+    resp_json = json.loads(resp.text)
+    LOG.log(LOG_SPAM, "total_count=%s", resp_json["total_count"])
+    return resp_json
+
+
+def _triage(df_report, search_nix_prs):
+    LOG.debug("")
+    df = df_report.copy()
+    uids = ["vuln_id", "package", "version", "url", "sortcol"]
+    if "whitelist" in df.columns:
+        uids.append("whitelist")
+        uids.append("whitelist_comment")
+    df_vuln_pkgs = df.groupby(by=uids).size().reset_index(name="count")
+    LOG.debug("Number of vulnerable packages: %s", df_vuln_pkgs.shape[0])
+    if df_vuln_pkgs.empty:
+        return df_vuln_pkgs
+    df_log(df_vuln_pkgs, LOG_SPAM)
+    # Find the repology version info for vulnerable packages
+    df_vuln_pkgs = _query_repology_versions(df_vuln_pkgs)
+    LOG.debug("Vulnerable pkgs with repology version info: %s", df_vuln_pkgs.shape[0])
+    df_log(df_vuln_pkgs, LOG_SPAM)
+    # Classify each vulnerable package
+    df_vuln_pkgs["classify"] = df_vuln_pkgs.apply(_vuln_update_classify, axis=1)
+    # Find potentially relevant nixpkgs PR
+    if search_nix_prs:
+        LOG.info("Querying nixpkgs github PRs")
+        df_vuln_pkgs["nixpkgs_pr"] = df_vuln_pkgs.apply(_vuln_nixpkgs_pr, axis=1)
+    # Sort the data based on the following columns
+    sort_cols = ["sortcol", "package", "version_local"]
+    df_vuln_pkgs.sort_values(by=sort_cols, ascending=False, inplace=True)
+    return df_vuln_pkgs
+
+
+################################################################################
+
+# Helpers
 
 
 def _reformat_scanner(val):
@@ -388,7 +755,7 @@ def _exit_unless_command_exists(name):
 def load_whitelist(whitelist_csv_path):
     """
     Load vulnerability whitelist from the given path. Returns None
-    if the whitelist not a valid vulnerability whitelist. Otherwise
+    if the whitelist is not a valid vulnerability whitelist. Otherwise
     returns whitelist_csv_path as dataframe.
     """
     try:
@@ -471,6 +838,8 @@ def df_drop_whitelisted(df):
 
 ################################################################################
 
+# Main
+
 
 def main():
     """main entry point"""
@@ -504,8 +873,6 @@ def main():
     scanner.scan_osv(sbom_cdx_path)
     scanner.report(args, sbom_csv_path)
 
-
-################################################################################
 
 if __name__ == "__main__":
     main()
